requirements.txt
requests
pillow
pydantic
requests
enkanetwork.py
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
enkanetwork.py==1.4.5
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
pillow==11.3.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.32.4
typing-inspection==0.4.1
typing_extensions==4.14.1
yarl==1.20.1
enkacard